+++ components/pages/users/Create.vue
@@ -1,5 +1,7 @@
 <template>
   <FormTemplate>
+    <Label class-name="u-d-block u-mt3">{{ $t('form.name') }}</Label>
+    <Label class-name="u-d-block u-ml2 u-mt2">{{ form.name }}</Label>
     <Label class-name="u-d-block u-mt3">{{ $t('form.email') }}</Label>
     <Label class-name="u-d-block u-ml2 u-mt2">{{ form.email }}</Label>
     <Label class-name="u-d-block u-mt3">{{ $t('form.password') }}</Label>
@@ -27,10 +29,10 @@
     </div>
     <div class="u-mt10">
       <BtnWarning
-        name="passwordResetPre"
+        name="createPre"
         class-name="u-d-block u-w5 u-h1 u-mx-auto u-align-center u-text-sz4"
-        @click="passwordResetPre"
-        >パスワード変更</BtnWarning
+        @click="createPre"
+        >新規登録</BtnWarning
       >
     </div>
     <portal to="target">
@@ -61,23 +63,24 @@ export default {
     return {
       form: {
         email: this.$store.state.users.users.response.email,
+        name: this.$store.state.users.users.response.name,
         password: '',
         passwordConfirm: '',
       },
     }
   },
   methods: {
-    async passwordResetPre() {
+    async createPre() {
       this.$v.$touch()
       if (!this.$v.$invalid) {
-        await this.$store.dispatch('users/passwordReset', {
+        await this.$store.dispatch('users/create', {
           email: this.form.email,
           password: this.form.password,
         })
         const users = this.$store.state.users.users
-        if (users && typeof users.response !== 'undefined') {
+        if (users && users.success && typeof users.response !== 'undefined') {
           if (users.success) {
-            this.$router.push({ path: '/' })
+            this.$router.push({ path: '/users/login' })
           } else {
             this.alertDangerShow(users.response.msg)
           }
