+++ components/pages/users/Login.vue
@@ -1,16 +1,6 @@
 <template>
   <FormTemplate>
-    <Label class-name="u-d-block">{{ $t('form.name') }}</Label>
-    <TextPlane
-      class-name="u-d-block u-px2 u-py1 u-mt2 u-text-sz4"
-      :value="$v.form.name.$model"
-      placeholder="浦島 太郎"
-      @input="$v.form.name.$model = $event"
-    ></TextPlane>
-    <div v-if="$v.form.name.$error && !$v.form.name.required" class="u-mt2 u-text-danger">
-      {{ $t('form.name') }}{{ $t('error.required') }}
-    </div>
-    <Label class-name="u-d-block u-mt3">{{ $t('form.email') }}</Label>
+    <Label class-name="u-d-block">{{ $t('form.email') }}</Label>
     <TextEmail
       class-name="u-d-block u-px2 u-py1 u-mt2 u-text-sz4"
       name="email"
@@ -22,25 +12,23 @@
       {{ $t('form.email') }}{{ $t('error.required') }}
     </div>
     <div v-if="$v.form.email.$error && !$v.form.email.email" class="u-mt2 u-text-danger">{{ $t('error.email') }}</div>
-    <Label class-name="u-d-block u-mt3">{{ $t('form.emailConfirm') }}</Label>
-    <TextEmail
+    <Label class-name="u-d-block u-mt3">{{ $t('form.password') }}</Label>
+    <TextPassword
       class-name="u-d-block u-px2 u-py1 u-mt2 u-text-sz4"
-      name="email"
-      :value="$v.form.emailConfirm.$model"
-      @input="$v.form.emailConfirm.$model = $event"
-    ></TextEmail>
-    <div v-if="$v.form.emailConfirm.$error && !$v.form.emailConfirm.required" class="u-mt2 u-text-danger">
-      {{ $t('form.emailConfirm') }}{{ $t('error.required') }}
+      name="password"
+      :value="$v.form.password.$model"
+      @input="$v.form.password.$model = $event"
+    ></TextPassword>
+    <div v-if="$v.form.password.$error && !$v.form.password.required" class="u-mt2 u-text-danger">
+      {{ $t('form.password') }}{{ $t('error.required') }}
     </div>
-    <div v-if="$v.form.emailConfirm.$error && !$v.form.emailConfirm.sameAsEmail" class="u-mt2 u-text-danger">
-      {{ $t('form.email') }}と{{ $t('form.emailConfirm') }}{{ $t('error.sameAs') }}
-    </div>
+    <Link href="passwordResetPre" class-name="u-flex-end u-mt3">パスワードを忘れた</Link>
     <div class="u-mt10">
-      <BtnWarning
-        name="createPre"
-        class-name="u-d-block u-w5 u-h1 u-mx-auto u-align-center u-text-sz4"
-        @click="createPre"
-        >新規登録</BtnWarning
+      <BtnPrimary name="login" class-name="u-d-block u-w5 u-h1 u-mx-auto u-align-center u-text-sz4" @click="login"
+        >ログイン</BtnPrimary
+      >
+      <LinkBtnWarning href-name="createPre" class-name="u-d-block u-w5 u-h1 u-mt2 u-mx-auto u-align-center u-text-sz4"
+        >新規登録</LinkBtnWarning
       >
     </div>
     <portal to="target">
@@ -57,39 +45,65 @@
 </template>
 
 <script>
-import { required, email, sameAs } from 'vuelidate/lib/validators'
+import { mapGetters } from 'vuex'
+import { required, email } from 'vuelidate/lib/validators'
 import FormTemplate from '@/components/templates/FormTemplate'
 import Label from '@/components/atoms/Label'
-import TextPlane from '@/components/atoms/TextPlane'
 import TextEmail from '@/components/atoms/TextEmail'
-import BtnWarning from '@/components/atoms/BtnWarning'
+import TextPassword from '@/components/atoms/TextPassword'
+import Link from '@/components/atoms/Link'
+import BtnPrimary from '@/components/atoms/BtnPrimary'
+import LinkBtnWarning from '@/components/atoms/LinkBtnWarning'
 import AlertInfo from '@/components/molecules/dialogs/AlertInfo'
 import AlertDanger from '@/components/molecules/dialogs/AlertDanger'
 
 export default {
-  components: { FormTemplate, Label, TextPlane, TextEmail, BtnWarning, AlertInfo, AlertDanger },
+  components: {
+    FormTemplate,
+    Label,
+    TextEmail,
+    TextPassword,
+    Link,
+    BtnPrimary,
+    LinkBtnWarning,
+    AlertInfo,
+    AlertDanger,
+  },
   data() {
     return {
       form: {
-        name: '',
         email: '',
-        emailConfirm: '',
+        password: '',
       },
     }
   },
+  computed: {
+    ...mapGetters(['users']),
+  },
+  mounted() {
+    this.$nextTick(() => {
+      const users = this.$store.state.users.users
+      if (users && users.success && typeof users.response !== 'undefined') {
+        if (users.success) {
+          this.alertInfoShow(users.response.msg)
+        } else {
+          this.alertDangerShow(users.response.msg)
+        }
+      }
+    })
+  },
   methods: {
-    async createPre() {
+    async login() {
       this.$v.$touch()
       if (!this.$v.$invalid) {
-        await this.$store.dispatch('users/createPre', {
-          name: this.form.name,
+        await this.$store.dispatch('users/login', {
           email: this.form.email,
+          password: this.form.password,
         })
         const users = this.$store.state.users.users
-        if (users && typeof users.response !== 'undefined') {
+        if (users && users.success && typeof users.response !== 'undefined') {
           if (users.success) {
-            this.alertInfoShow(users.response.msg)
-            this.clear()
+            console.log(users)
           } else {
             this.alertDangerShow(users.response.msg)
           }
@@ -106,18 +120,11 @@ export default {
     alertDangerShow(msg) {
       this.$refs.alertDanger.show(msg)
     },
-    clear() {
-      this.form.name = ''
-      this.form.email = ''
-      this.form.emailConfirm = ''
-      this.$v.$reset()
-    },
   },
   validations: {
     form: {
-      name: { required },
       email: { required, email },
-      emailConfirm: { required, sameAsEmail: sameAs('email') },
+      password: { required },
     },
   },
 }
