+++ components/pages/users/PasswordResetPre.vue
@@ -12,23 +12,12 @@
       {{ $t('form.email') }}{{ $t('error.required') }}
     </div>
     <div v-if="$v.form.email.$error && !$v.form.email.email" class="u-mt2 u-text-danger">{{ $t('error.email') }}</div>
-    <Label class-name="u-d-block u-mt3">{{ $t('form.password') }}</Label>
-    <TextPassword
-      class-name="u-d-block u-px2 u-py1 u-mt2 u-text-sz4"
-      name="password"
-      :value="$v.form.password.$model"
-      @input="$v.form.password.$model = $event"
-    ></TextPassword>
-    <div v-if="$v.form.password.$error && !$v.form.password.required" class="u-mt2 u-text-danger">
-      {{ $t('form.password') }}{{ $t('error.required') }}
-    </div>
-    <Link href="passwordResetPre" class-name="u-flex-end u-mt3">パスワードを忘れた</Link>
     <div class="u-mt10">
-      <BtnPrimary name="login" class-name="u-d-block u-w5 u-h1 u-mx-auto u-align-center u-text-sz4" @click="login"
-        >ログイン</BtnPrimary
-      >
-      <LinkBtnWarning href-name="/createPre" class-name="u-d-block u-w5 u-h1 u-mt2 u-mx-auto u-align-center u-text-sz4"
-        >新規登録</LinkBtnWarning
+      <BtnWarning
+        name="passwordResetPre"
+        class-name="u-d-block u-w5 u-h1 u-mx-auto u-align-center u-text-sz4"
+        @click="passwordResetPre"
+        >パスワード変更</BtnWarning
       >
     </div>
     <portal to="target">
@@ -45,65 +34,34 @@
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
 import { required, email } from 'vuelidate/lib/validators'
 import FormTemplate from '@/components/templates/FormTemplate'
 import Label from '@/components/atoms/Label'
 import TextEmail from '@/components/atoms/TextEmail'
-import TextPassword from '@/components/atoms/TextPassword'
-import Link from '@/components/atoms/Link'
-import BtnPrimary from '@/components/atoms/BtnPrimary'
-import LinkBtnWarning from '@/components/atoms/LinkBtnWarning'
+import BtnWarning from '@/components/atoms/BtnWarning'
 import AlertInfo from '@/components/molecules/dialogs/AlertInfo'
 import AlertDanger from '@/components/molecules/dialogs/AlertDanger'
 
 export default {
-  components: {
-    FormTemplate,
-    Label,
-    TextEmail,
-    TextPassword,
-    Link,
-    BtnPrimary,
-    LinkBtnWarning,
-    AlertInfo,
-    AlertDanger,
-  },
+  components: { FormTemplate, Label, TextEmail, BtnWarning, AlertInfo, AlertDanger },
   data() {
     return {
       form: {
         email: '',
-        password: '',
       },
     }
   },
-  computed: {
-    ...mapGetters(['users']),
-  },
-  mounted() {
-    this.$nextTick(() => {
-      const users = this.$store.state.users.users
-      if (users && typeof users.response !== 'undefined') {
-        if (users.success) {
-          this.alertInfoShow(users.response.msg)
-        } else {
-          this.alertDangerShow(users.response.msg)
-        }
-      }
-    })
-  },
   methods: {
-    async login() {
+    async passwordResetPre() {
       this.$v.$touch()
       if (!this.$v.$invalid) {
-        await this.$store.dispatch('users/login', {
+        await this.$store.dispatch('users/passwordResetPre', {
           email: this.form.email,
-          password: this.form.password,
         })
         const users = this.$store.state.users.users
-        if (users && typeof users.response !== 'undefined') {
+        if (users && users.success && typeof users.response !== 'undefined') {
           if (users.success) {
-            console.log(users)
+            this.alertInfoShow(users.response.msg)
           } else {
             this.alertDangerShow(users.response.msg)
           }
@@ -124,7 +82,6 @@ export default {
   validations: {
     form: {
       email: { required, email },
-      password: { required },
     },
   },
 }
