+++ pages/users/createPre.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import CreatePre from '@/components/CreatePre.vue'
+import CreatePre from '@/components/pages/users/CreatePre.vue'
 export default Vue.extend({
   name: 'CreatePrePage',
   components: { CreatePre },
