+++ pages/users/passwordReset.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import PasswordReset from '@/components/PasswordReset.vue'
+import PasswordReset from '@/components/pages/users/PasswordReset.vue'
 
 export default Vue.extend({
   name: 'PasswordResetPage',
